move unit tests to new module, Edges derie macro debugging
