+++ crates/apecs-derive/Cargo.toml
@@ -16,7 +16,7 @@ proc-macro = true
 bench = false
 
 [dependencies]
-apecs-derive-canfetch = { path = "../apecs-derive-canfetch", version = "^0.2.3" }
-proc-macro2 = "^1.0"
-quote = "^1.0"
-syn = "^1.0"
+moongraph-macros-syntax = { path = "../../../moongraph/crates/moongraph-macros-syntax" }
+proc-macro2 = "1.0"
+quote = "1.0"
+syn = { version = "2.0", features = ["full"] }
