+++ crates/apecs/Cargo.toml
@@ -26,7 +26,7 @@ bench = false
 any_vec = "0.10"
 apecs-derive = { version = "0.2.3", path = "../apecs-derive" }
 async-channel = "1.6"
-moongraph = { version = "0.4", path = "../../../moongraph/crates/moongraph" }
+moongraph = { version = "0.4", path = "../../../moongraph/crates/moongraph", default-features = false, features = ["parallel"] }
 itertools = "0.10"
 log = "0.4"
 parking_lot = "0.12"
